requirements.txt
arrow
discord.py
tinydb
fuzzywuzzy
pyowm
tungsten
wikia
imgurpython
cleverbot
urbandict
langdetect
iso639
flask
markovify
heroku3